requirements.txt
Faker==40.12.0
greenlet==3.3.2
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
